Revert "feat: load and merge all descriptors (#1271)"
This reverts commit c9b51183e9cbc32a4dcc1c025a37d66c97d46047.
