+++ akka-javasdk-annotation-processor-tests/actions-descriptors/src/main/java/com/example/AbstracInbetweenAction.java
@@ -6,4 +6,4 @@ package com.example;
 
 import akka.javasdk.timedaction.TimedAction;
 
-public abstract class AbstractInBetweenAction extends TimedAction {}
+public abstract class AbstracInbetweenAction extends TimedAction {}
