+++ akka-javasdk-annotation-processor-tests/actions-descriptors/src/main/java/com/example/HierarchyTimedAction.java
@@ -7,4 +7,4 @@ package com.example;
 import akka.javasdk.annotations.Component;
 
 @Component(id = "hierarchy-timed-action")
-public class HierarchyTimedAction extends AbstractInBetweenAction {}
+public class HierarchyTimedAction extends AbstracInbetweenAction {}
