+++ akka-javasdk-annotation-processor-tests/entites-and-views-descriptors/src/test/scala/com/example/EntitiesAndViewsDescriptorSpec.scala
@@ -12,7 +12,7 @@ class EntitiesAndViewsDescriptorSpec extends AnyWordSpec with Matchers {
 
   "akka-javasdk-components.conf" should {
     "have correct configuration" in {
-      val config = ConfigFactory.load("META-INF/akka-javasdk-components_com-example_test.conf")
+      val config = ConfigFactory.load("META-INF/akka-javasdk-components.conf")
 
       val keyValueComponents = config.getStringList("akka.javasdk.components.key-value-entity")
       keyValueComponents.size() shouldBe 2
