+++ akka-javasdk-tests/src/test/resources/META-INF/akka-javasdk-components.conf
